Move AoS LCAO classes to MolecularOrbitals
