+++ src/QMCWaveFunctions/Fermion/SlaterDetBuilder.cpp
@@ -37,8 +37,8 @@
 #include "QMCWaveFunctions/Fermion/RNDiracDeterminantBaseAlternate.h"
 //Cannot use complex with SlaterDetOpt
 #include "QMCWaveFunctions/MolecularOrbitals/NGOBuilder.h"
-#include "QMCWaveFunctions/LocalizedBasisSet.h"
-#include "QMCWaveFunctions/LCOrbitalSetOpt.h"
+#include "QMCWaveFunctions/MolecularOrbitals/LocalizedBasisSet.h"
+#include "QMCWaveFunctions/MolecularOrbitals/LCOrbitalSetOpt.h"
 #include "QMCWaveFunctions/Fermion/SlaterDetOpt.h"
 #endif
 #ifdef QMC_CUDA
